requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
Django==3.2.8
django-environ==0.8.0
djangorestframework==3.12.4
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
